pyproject.toml
[project]
name = "sf-lambda-functions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.5",
    "polars>=1.32.3",
    "polars-ols>=0.3.5",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
]
